refactor(admin): decompose 1095-line monolith into 7 concern modules
Before: admin.py (1095 lines, 23 route handlers, 7 concerns)
After: 8 focused modules, largest 241 lines, all under 400

Files:
  _admin_helpers.py    (39 lines)  shared audit + _client_ip
  admin_players.py     (237 lines) §3.2 Player management
  admin_games.py       (207 lines) §3.3 Game inspection
  admin_system.py      (118 lines) §3.4 Health / metrics / log / purge
  admin_moderation.py  (148 lines) §3.5 Moderation queue
  admin_rate_limits.py (125 lines) §3.6 Rate-limit management
  admin_operations.py  (241 lines) §3.7-3.8 Audit / jobs / config
  admin_prompts.py     (142 lines) §3.8 Prompt management (prior PR)
  admin.py             (40 lines)  Re-export hub

All 25 /admin routes verified intact via create_app() smoke test.
